requirements.txt
cffi==1.16.0
psutil==5.9.8
pycparser==2.21
pycryptodome==3.20.0
